# docs/swagger.yaml
basePath: /api
definitions:
...
  groups.Group:
    properties:
      created_at:
        type: string
      id:
        type: string
      members:
        items:
          type: string
        type: array
      name:
        type: string
    type: object
  groups.JoinRequest:
    properties:
      created_at:
        type: string
      group_id:
        type: string
      id:
        type: string
      user_id:
        type: string
    type: object
  groups.createGroupReq:
    properties:
      members:
        items:
          type: string
        type: array
      name:
        type: string
    required:
    - name
    type: object
  messages.Message:
    properties:
      chat_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      sender_id:
        type: string
    type: object
  messages.sendReq:
    properties:
      content:
        type: string
    required:
    - content
    type: object
...
  posts.Post:
    properties:
      author_id:
        type: integer
      category:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      likes_count:
        type: integer
      saved_count:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    required:
    - category
    type: object
host: localhost:8080
info:
  contact:
    email: support@uniconnect.local
    name: API Support
  description: API for University Social Platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://uniconnect.local/terms/
  title: UniConnect API
  version: "1.0"
paths:
  /admin/groups:
    post:
      consumes:
      - application/json
      description: Admin creates a new group
      parameters:
      - description: Group info
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/groups.createGroupReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/groups.Group'
        "400":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a group
      tags:
      - Groups
  /api/messages/{chatId}:
    get:
      description: Returns all messages in a chat
      parameters:
      - description: Chat ID
        in: path
        name: chatId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/messages.Message'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List chat messages
      tags:
      - Messages
    post:
      consumes:
      - application/json
      description: Send a message to a chat
      parameters:
      - description: Chat ID
        in: path
        name: chatId
        required: true
        type: string
      - description: Message content
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/messages.sendReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/messages.Message'
        "400":
          description: error
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Send a message
      tags:
      - Messages
...
  /groups/:
    get:
      description: List groups (user sees only joined groups)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/groups.Group'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List groups
      tags:
      - Groups
  /groups/{groupId}/join:
    post:
      description: Student sends join request
      parameters:
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/groups.JoinRequest'
        "401":
          description: unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Request to join a group
      tags:
      - Groups
  /posts:
    get:
      description: Returns all posts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/posts.Post'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List posts
      tags:
      - Posts
    post:
      consumes:
      - application/json
      description: Authenticated user creates a post
      parameters:
      - description: Post info
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/posts.Post'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/posts.Post'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a post
      tags:
      - Posts
  /posts/{id}:
    delete:
      description: Deletes a post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a post
      tags:
      - Posts
    put:
      consumes:
      - application/json
      description: Update a post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Post data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/posts.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posts.Post'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a post
      tags:
      - Posts
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"